requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
anthropic==0.40.0
sqlalchemy==2.0.35
aiosqlite==0.20.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.5.2
groq>=0.11.0
pyjwt>=2.8.0
